feat: 사용자 장소 타입 수정 로직 변경(#100)

+++ src/main/java/org/example/tokpik_be/type/service/PlaceTypeService.java
@@ -57,14 +57,16 @@ public class PlaceTypeService {
             throw new GeneralException(TypeException.DUPLICATE_TYPES);
         }
 
+        List<PlaceType> placeTypes = placeTypeRepository.findAllById(request.placeTypeIds());
+
+        if (placeTypes.size() != request.placeTypeIds().size()) {
+            throw new GeneralException(TypeException.TYPE_NOT_FOUND);
+        }
+
         userPlaceTypeRepository.deleteByUserId(user.getId());
 
-        for (long typeId : request.placeTypeIds()) {
-            if (!placeTypeRepository.existsById(typeId)) {
-                throw new GeneralException(TypeException.TYPE_NOT_FOUND);
-            }
-            UserPlaceType userPlaceType = new UserPlaceType(user.getId(),
-                placeTypeRepository.findById(typeId).get());
+        for (PlaceType placeType : placeTypes) {
+            UserPlaceType userPlaceType = new UserPlaceType(user.getId(), placeType);
             userPlaceTypeRepository.save(userPlaceType);
         }
